requirements.txt
aggdraw==1.3.18.post0
colorama==0.4.6
greenlet==3.0.3
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
playwright==1.44.0
pyee==11.1.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
six==1.16.0
typing_extensions==4.12.0
tzdata==2024.1
